+++ fxmanifest.lua
@@ -1,5 +1,6 @@
 fx_version 'cerulean'
 game 'gta5'
+lua54 'yes'
 
 name 'pxCommands'
 author 'CodeMeAPixel'
@@ -7,7 +8,7 @@ contact 'hey@codemeapixel.dev'
 description 'A modular and extendable chat/command system'
 license 'GPL-3.0-or-later'
 repository 'https://github.com/CodeMeAPixel/pxCommands'
-version '0.1.0'
+version '0.2.0'
 
 shared_scripts {
     'system/config.lua'