+++ modules/cl_overhead_text.lua
@@ -7,6 +7,9 @@ local FloatingTextMaxDistance = 50.0
 
 RegisterNetEvent("pxc:showFloatingText")
 AddEventHandler("pxc:showFloatingText", function(target, text)
+    if source ~= '' and tonumber(source) ~= nil then
+        return
+    end
     table.insert(FloatingText, {target = GetPlayerFromServerId(target), text = text, life = GetGameTimer() + FloatingTextDuration})
 end)
 
@@ -37,18 +40,23 @@ end
 
 Citizen.CreateThread(function()
     while true do
-        Wait(0)
-        local lpos = GetEntityCoords(PlayerPedId())
-        for n, text in next, FloatingText do
-            if NetworkIsPlayerActive(text.target) then
-                local ply = GetPlayerPed(text.target)
-                local pos = GetWorldPositionOfEntityBone(ply, GetEntityBoneIndexByName(ply, "head"))
-                if #(lpos - pos) <= FloatingTextMaxDistance then
-                    DrawText3D(text.text, pos.x, pos.y, pos.z + 0.2)
+        if #FloatingText == 0 then
+            Wait(500)
+        else
+            Wait(0)
+            local now = GetGameTimer()
+            local lpos = GetEntityCoords(PlayerPedId())
+            for n = #FloatingText, 1, -1 do
+                local text = FloatingText[n]
+                if text.life < now then
+                    table.remove(FloatingText, n)
+                elseif NetworkIsPlayerActive(text.target) then
+                    local ply = GetPlayerPed(text.target)
+                    local pos = GetWorldPositionOfEntityBone(ply, GetEntityBoneIndexByName(ply, "head"))
+                    if #(lpos - pos) <= FloatingTextMaxDistance then
+                        DrawText3D(text.text, pos.x, pos.y, pos.z + 0.2)
+                    end
                 end
-            end
-            if text.life < GetGameTimer() then
-                table.remove(FloatingText, n)
             end
         end
     end