+++ system/config.lua
@@ -8,6 +8,7 @@ Config.CheckUpdates = true
 
 Config.Callbacks = {
     onCommandExecuted = nil,
+    onCommandFailed = nil,
 }
 
 Config.AdminCheck = nil
@@ -15,4 +16,14 @@ Config.AdminCheck = nil
 Config.Formatting = {
     showPlayerId = true,
     useFrameworkName = true,
+}
+
+Config.Cooldowns = {
+    default = 0,
+}
+
+Config.Webhook = {
+    enabled = false,
+    url = "",
+    handler = nil,
 }